+++ src/template_nn/networks/cnn.py
@@ -7,7 +7,7 @@ from ..args_val import is_iterable, is_positive_int
 from .base_nn import BaseNetwork
 from .cml import CML
 from .fnn import FNN
-from .nn_keys import MODEL_KEYS
+from ..nn_keys import MODEL_KEYS
 
 
 class CNN(BaseNetwork):
